Prioritize static paths in infect
The arguments to infect are reversed so that the earlier arguments show
up earlier in the load path.  This results in a lot of unintuitive
behaviors and I generally consider it to have been a mistake.  To
mitigate this, try pulling out the static, absolute paths and applying
them first.  You almost always want this, as it enables the expected
nesting in `pathogen#infect('~/Code/vim/', 'bundle/{}')`.