Make a note for users of .dotfiles repos
When setting up a [dotfiles](https://dotfiles.github.io/) repo, it is typical for the entire contents of the `.vim` directory to be added into the `.dotfiles` git repo. Having the pathogen plugins clones as nested repositories is problematic in this case and we need to instead set them up as git submodules.